fixing login and logout api

+++ frontend/src/hooks/useAxiosPrivate.jsx
@@ -3,36 +3,30 @@ import axios from "axios";
 const URL = import.meta.env.VITE_API_END_POINT;
 
 const useAxiosPrivate = () => {
-
-    const Token = localStorage.getItem("token");
-
     const axiosPrivate = axios.create({
         baseURL: URL,
-        headers: {
-            Authorization: `Bearer ${Token}`,
-        },
     });
 
     useEffect(() => {
         const requestIntercept = axiosPrivate.interceptors.request.use(
             (config) => {
-                if (!config.headers.Authorization) {
-                    config.headers.Authorization = `Bearer ${accessToken}`;
+                const token = JSON.parse(localStorage.getItem("token"));
+                console.log(token);
+                if (token) {
+                    config.headers.Authorization = `Token ${token}`;
                 }
                 return config;
             },
-
             (error) => Promise.reject(error)
         );
 
         const responseIntercept = axiosPrivate.interceptors.response.use(
             (response) => response,
-            async (error) => {
-
-                if (error.response.status === 401) {
-                    localStorage.clear()
+            (error) => {
+                if (error.response && error.response.status === 401) {
+                    // Handle 401 unauthorized (e.g., clear local storage or redirect to login)
+                    localStorage.clear();
                 }
-
                 return Promise.reject(error);
             }
         );
@@ -41,7 +35,7 @@ const useAxiosPrivate = () => {
             axiosPrivate.interceptors.request.eject(requestIntercept);
             axiosPrivate.interceptors.response.eject(responseIntercept);
         };
-    }, [Token, axiosPrivate]);
+    }, [axiosPrivate]);
 
     return axiosPrivate;
 };
